Refactor logic for handling unmanaged Java upgrade
Previously we were just not attempting to upgrade Java if it wasn't
"managed", i.e. matching the scijava.app.java-root. However, old Fiji
installs were erronously showing as "unmanaged" because we are changing
the /java/ subdir and the old bundled Java wasn't being recognized.

So now the current logic is:

- If not in a .app subdir (definitely unmanaged), offer to manage
- If bundled, managed, or unmanaged but opted-in, we first check for an
  already-present "good enough" install and try switching to that
- If no "good enough" install then we see if we can find a remote Java
  to download using scijava.app.java-links. If found for this platform
  we offer to download and install it.

The user can opt-out of the management offer and the upgrade offer.